Added Tool Bar on top
Removed the temporary sign out button on the user's profile and added it to the tool bar.

There is now a Drawer Navigation that can be swiped to the right to show its content. The buttons there currently has no functions.

FIX: The layout of the fragments are not properly fixed